:a fixes conflicts

+++ BuzzyMail/BuzzyMail/ReplyMailViewController.swift
@@ -13,6 +13,7 @@ class ReplyMailViewController: UIViewController {
 
     let service = OutlookService.shared()
     var newEmail:Message?
+    var isNewMail:Bool = false
     var container: MailContentTableViewController?
 
     let dispatchGroup = DispatchGroup()
@@ -24,33 +25,87 @@ class ReplyMailViewController: UIViewController {
     }
 
     @IBAction func sendButtonPressed(_ sender: Any) {
-
+        
+        if (newEmail == nil) {
+            self.newEmail = Message()
+            self.isNewMail = true
+        }
         self.dispatchGroup.enter()
         if let richTextEditor = container?.richTextEditor {
-            self.newEmail?.body!.content = richTextEditor.updatedText!
+            if self.newEmail?.body == nil {
+                self.newEmail?.body = Body(contentType: "html", content: richTextEditor.updatedText!)
+            }
+            else {
+                self.newEmail?.body?.content = richTextEditor.updatedText!
+            }
             self.newEmail?.subject = container!.subjectTextField.text!
             
             self.newEmail?.toRecipients = [EmailAddresses]()
             let toTextFieldArray = container!.toTextField.text!.components(separatedBy: ", ")
             for toTextField in toTextFieldArray {
-                self.newEmail?.toRecipients!.append(EmailAddresses(emailAddress: EmailAddress(name: "", address: toTextField)))
+                self.newEmail?.toRecipients?.append(EmailAddresses(emailAddress: EmailAddress(name: "", address: toTextField)))
             }
             
             self.newEmail?.ccRecipients = [EmailAddresses]()
             let ccTextFieldArray = container!.ccTextField.text!.components(separatedBy: ", ")
             for ccTextField in ccTextFieldArray {
-                self.newEmail?.ccRecipients!.append(EmailAddresses(emailAddress: EmailAddress(name: "", address: ccTextField)))
+                self.newEmail?.ccRecipients?.append(EmailAddresses(emailAddress: EmailAddress(name: "", address: ccTextField)))
             }
             self.dispatchGroup.leave()
         }
         self.dispatchGroup.notify(queue: .main) {
             self.dispatchGroup2.enter()
-            self.updateReply()
+            if self.isNewMail {
+                self.createMail()
+            }
+            else {
+                self.updateReply()
+            }
             self.dispatchGroup2.notify(queue: .main) {
                 self.sendReply()
                 self.performSegue(withIdentifier: "closeDraft", sender: sender)
             }
         }
+
+        /*
+        else {
+            var toRecipientsList = [EmailAddresses]()
+            for row in message["toRecipients"].arrayValue {
+                toRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                  address: row["emailAddress"]["address"].stringValue)))
+            }
+            
+            var ccRecipientsList = [EmailAddresses]()
+            for row in message["ccRecipients"].arrayValue {
+                ccRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                  address: row["emailAddress"]["address"].stringValue)))
+            }
+            
+            var bccRecipientsList = [EmailAddresses]()
+            for row in message["bccRecipients"].arrayValue {
+                bccRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                   address: row["emailAddress"]["address"].stringValue)))
+            }
+            
+            //receivedDateTime: Formatter.dateToString(date: message["receivedDateTime"]),
+            let newMsg = Message(
+                id: message["id"].stringValue,
+                receivedDateTime: message["receivedDateTime"].stringValue,
+                hasAttachments: message["hasAttachments"].boolValue,
+                subject: message["subject"].stringValue,
+                bodyPreview: message["bodyPreview"].stringValue,
+                isRead: message["isRead"].boolValue,
+                isDraft: message["isDraft"].boolValue,
+                body: Body(contentType: message["body"]["contentType"].stringValue,
+                           content: message["body"]["content"].stringValue),
+                from: EmailAddresses(emailAddress: EmailAddress(name: message["from"]["emailAddress"]["name"].stringValue,
+                                                                address: message["from"]["emailAddress"]["address"].stringValue)),
+                toRecipients: toRecipientsList,
+                ccRecipients: ccRecipientsList,
+                bccRecipients: bccRecipientsList)
+
+        }
+ */
     }
 
 
@@ -66,19 +121,26 @@ class ReplyMailViewController: UIViewController {
     @IBAction func cancelButtonPressed(_ sender: Any) {
 
         let deleteDraftActionHandler = { (action:UIAlertAction!) -> Void in
-            self.service.deleteMessage(message: self.newEmail!) {_ in }
+            if (self.newEmail != nil) {
+                self.service.deleteMessage(message: self.newEmail!) {_ in }
+            }
             self.performSegue(withIdentifier: "closeDraft", sender: action)
         }
 
         let saveDraftActionHandler = { (action:UIAlertAction!) -> Void in
-            self.dispatchGroup3.enter()
-            if let richTextEditor = self.container?.richTextEditor {
-                self.newEmail?.body!.content = richTextEditor.updatedText!
-                self.newEmail?.subject = self.container!.subjectTextField.text!
-                self.dispatchGroup3.leave()
+            if (self.newEmail != nil) {
+                self.dispatchGroup3.enter()
+                if let richTextEditor = self.container?.richTextEditor {
+                    self.newEmail?.body!.content = richTextEditor.updatedText!
+                    self.newEmail?.subject = self.container!.subjectTextField.text!
+                    self.dispatchGroup3.leave()
+                }
+                self.dispatchGroup3.notify(queue: .main) {
+                    self.service.updateReply(message: self.newEmail!) {_ in }
+                    self.performSegue(withIdentifier: "closeDraft", sender: action)
+                }
             }
-            self.dispatchGroup3.notify(queue: .main) {
-                self.service.updateReply(message: self.newEmail!) {_ in }
+            else {
                 self.performSegue(withIdentifier: "closeDraft", sender: action)
             }
         }
@@ -96,6 +158,58 @@ class ReplyMailViewController: UIViewController {
     }
 
 
+    func createMail() {
+        NSLog("createMail called")
+        service.createMail(message: newEmail!) {
+            message in
+            if let message = message {
+                NSLog("Create Mail Success")
+                NSLog("Pompernikkel3: " + message["subject"].stringValue)
+  
+                var toRecipientsList = [EmailAddresses]()
+                for row in message["toRecipients"].arrayValue {
+                    toRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                      address: row["emailAddress"]["address"].stringValue)))
+                }
+                
+                var ccRecipientsList = [EmailAddresses]()
+                for row in message["ccRecipients"].arrayValue {
+                    ccRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                      address: row["emailAddress"]["address"].stringValue)))
+                }
+                
+                var bccRecipientsList = [EmailAddresses]()
+                for row in message["bccRecipients"].arrayValue {
+                    bccRecipientsList.append(EmailAddresses(emailAddress: EmailAddress(name: row["emailAddress"]["name"].stringValue,
+                                                                                       address: row["emailAddress"]["address"].stringValue)))
+                }
+                
+                //receivedDateTime: Formatter.dateToString(date: message["receivedDateTime"]),
+                let newMsg = Message(
+                    id: message["id"].stringValue,
+                    receivedDateTime: message["receivedDateTime"].stringValue,
+                    hasAttachments: message["hasAttachments"].boolValue,
+                    subject: message["subject"].stringValue,
+                    bodyPreview: message["bodyPreview"].stringValue,
+                    isRead: message["isRead"].boolValue,
+                    isDraft: message["isDraft"].boolValue,
+                    body: Body(contentType: message["body"]["contentType"].stringValue,
+                               content: message["body"]["content"].stringValue),
+                    from: EmailAddresses(emailAddress: EmailAddress(name: message["from"]["emailAddress"]["name"].stringValue,
+                                                                    address: message["from"]["emailAddress"]["address"].stringValue)),
+                    toRecipients: toRecipientsList,
+                    ccRecipients: ccRecipientsList,
+                    bccRecipients: bccRecipientsList)
+                
+                self.newEmail = newMsg
+                self.dispatchGroup2.leave()
+            } else {
+                NSLog("Create Mail Fail")
+                self.dispatchGroup2.leave()
+            }
+        }
+    }
+    
 
     func sendReply() {
         NSLog("sendReply called")
